+++ include/phase4/engine/board/ordering/move_ordering.h
@@ -46,7 +46,7 @@ public:
 		}
 	}
 
-	static void assignQuietValues(const Session &session, const moves::Moves &moves, moves::MoveValues &moveValues, int32_t startIndex, int32_t ply) {
+	static void assignQuietValues(const Session &session, const moves::Moves &moves, moves::MoveValues &moveValues, size_t startIndex, int32_t ply) {
 		moveValues.resize(moves.size());
 		const Position &position = session.position();
 		for (size_t moveIndex = startIndex; moveIndex < moves.size(); ++moveIndex) {
